requirements.txt
python-telegram-bot==20.7
google-genai>=1.2.0
openai==1.12.0
